refactor: 開発用テストページを削除し、共有ユーティリティをlib/textbox-canvasに移動
テストページ7つ（dnd-kit-test, simple-dnd-kit-test, table-dnd-kit-test,
mathjax-offline-test, pdf-worker-offline-test, test/grid-layout-debug,
textbox-on-canvas-v4）を削除。textbox-on-canvas-v4の本番コードから参照
されていたutils/typesはsrc/lib/textbox-canvas/に移動し、全importパスを更新。
Navigationからdevナビゲーションを削除。
