--- src/app/textbox-on-canvas-v4/REFACTORING_REPORT.md
@@ -1,160 +1,0 @@
-# textbox-on-canvas-v4 リファクタリング完了レポート
-
-## 🎯 リファクタリング概要
-
-textbox-on-canvas-v4の958行の巨大なファイルを機能別に分割し、保守性・可読性・再利用性を大幅に向上させました。
-
-## 📊 リファクタリング前後の比較
-
-### ファイル構成の変化
-
-**リファクタリング前:**
-
-- TextboxCanvasPage.tsx: **958行** (単一の巨大ファイル)
-
-**リファクタリング後:**
-
-- TextboxCanvasPage.tsx: **277行** (-681行, 71%削減)
-- 4個のプレビューコンポーネント
-- 2個のカスタムフック
-- 統合ユーティリティファイル
-- 拡張型定義
-
-## 🏗️ 新しい構造
-
-### 1. コンポーネント分離 (`/components/`)
-
-```
-components/
-├── index.ts                    # 統合エクスポート
-├── TextboxPreview.tsx          # DIVプレビュー (51行)
-├── TextboxSvgPreview.tsx       # SVGプレビュー (76行)
-├── TextboxImagePreview.tsx     # Imageプレビュー (125行)
-└── TextboxCanvasPreview.tsx    # Canvasプレビュー (95行)
-```
-
-### 2. カスタムフック (`/hooks/`)
-
-```
-hooks/
-├── index.ts                    # 統合エクスポート
-├── useCanvasManagement.ts      # Canvas操作管理 (143行)
-└── useTextBoxOperations.ts     # テキストボックス操作 (164行)
-```
-
-### 3. ユーティリティ拡張 (`/utils/`)
-
-```
-utils/
-├── svgToCanvasUtils.ts         # SVG→Canvas統合ロジック (160行) ★新規
-├── textConversionUtils.ts      # テキスト変換 (523行)
-├── mathJaxUtils.ts            # MathJax処理 (435行)
-├── canvasUtils.ts             # Canvas描画 (283行)
-└── coordinateUtils.ts         # 座標計算 (184行)
-```
-
-### 4. 型定義強化 (`types.ts`)
-
-- 270行 (+139行)
-- UI Types、Rendering Types、Hook Types の分類
-- 新しい型：`RenderingStatus`、`SvgConversionOptions`、`CanvasManagementHook`等
-
-## ✨ 主要改善点
-
-### 1. **責任の分離**
-
-- **プレビューコンポーネント**: 各描画段階を独立したコンポーネントに分離
-- **カスタムフック**: Canvas管理とテキストボックス操作を分離
-- **統合ユーティリティ**: 重複するSVG変換ロジックを統一
-
-### 2. **可読性の向上**
-
-- メインファイル: 958行 → 277行 (71%削減)
-- 1ファイル1責任の原則を徹底
-- 機能別のファイル配置
-
-### 3. **再利用性の向上**
-
-- 共通SVG変換ロジックの統一化
-- プレビューコンポーネントの独立性
-- 型安全性の強化
-
-### 4. **保守性の向上**
-
-- バグの発生箇所を特定しやすい構造
-- 機能追加時の影響範囲を限定
-- テストコードの作成が容易
-
-## 🧩 統合されたロジック
-
-### SVG変換の統一化
-
-**新規作成: `svgToCanvasUtils.ts`**
-
-- `addMathJaxDefsToSvg()`: MathJax defs追加の統一関数
-- `generateSvgElement()`: SVG要素生成の統一関数
-- `createSvgBlobUrl()`: Blob URL作成の統一関数
-- `RenderingStatusManager`: 描画状態管理クラス
-
-## 🎨 プレビューシステムの改善
-
-4段階プレビューを完全に独立したコンポーネントに分離:
-
-1. **DIV プレビュー**: MathJax処理後のHTML表示
-2. **SVG プレビュー**: 図形データとして変換後の表示
-3. **Image プレビュー**: PNG変換後の表示
-4. **Canvas プレビュー**: Canvas描画結果の表示
-
-## 🔄 カスタムフックによる状態管理
-
-### `useCanvasManagement`
-
-- Canvas描画の統一管理
-- 背景画像サポート
-- エラーハンドリング強化
-
-### `useTextBoxOperations`
-
-- マウスイベント処理
-- テキストボックス選択・編集
-- ドラッグ＆ドロップ操作
-
-## 📈 パフォーマンスの改善
-
-- **メモ化**: 不要な再計算を防止
-- **状態管理の最適化**: 必要な時のみ再描画
-- **メモリ効率**: Blob URLの適切なクリーンアップ
-
-## 🎯 今後の拡張性
-
-### 容易な機能追加
-
-- 新しいプレビューモードの追加が简单
-- Canvas描画機能の拡張が容易
-- テキストフォーマット機能の追加が簡単
-
-### テスタビリティ向上
-
-- 個別コンポーネントのユニットテスト
-- カスタムフックのテスト
-- 統合テストの実装が容易
-
-## 🏆 技術的成果
-
-- **行数削減**: 958行 → 277行メイン + 分散ファイル群
-- **責任分離**: 1つの巨大ファイル → 機能別の小さなファイル群
-- **型安全性**: 139行の新規型定義追加
-- **保守性**: ファイル分割による改善
-- **可読性**: コメント・ドキュメントの充実
-
-## 📝 結論
-
-このリファクタリングにより、textbox-on-canvas-v4は：
-
-1. **保守しやすい**アーキテクチャに変更
-2. **拡張しやすい**設計に改善
-3. **理解しやすい**コード構造に変更
-4. **テストしやすい**モジュール化完了
-5. **パフォーマンス向上**の基盤構築
-
-プロダクションレベルの品質とスケーラビリティを持つモジュール設計が完成しました。
