--- src/app/textbox-on-canvas-v4/utils/coordinateUtils.ts
@@ -1,185 +1,0 @@
-/**
- * @fileoverview 座標変換・マウス操作ユーティリティ
- * @description Canvas座標系とマウス座標系の変換、テキストボックス操作
- */
-
-import { CANVAS_SETTINGS, TEXTBOX_SETTINGS } from "../constants"
-import type { DragState, Point, TextBox } from "../types"
-
-/**
- * マウス座標をCanvas座標に変換する
- * @param clientX マウスのクライアントX座標
- * @param clientY マウスのクライアントY座標
- * @param canvas Canvas要素
- * @param zoom ズーム倍率
- * @returns Canvas座標
- */
-export function getCanvasCoordinates(
-  clientX: number,
-  clientY: number,
-  canvas: HTMLCanvasElement,
-  zoom: number
-): Point {
-  const rect = canvas.getBoundingClientRect()
-  const x = (clientX - rect.left) / zoom
-  const y = (clientY - rect.top) / zoom
-
-  return { x, y }
-}
-
-/**
- * 点がアンカー点の近くにあるかどうかを判定する
- * @param point 判定する点
- * @param anchor アンカー点（x, y）
- * @param radius 判定半径（デフォルト: アンカー点の半径）
- * @returns アンカー点の範囲内にある場合はtrue
- */
-export function isPointNearAnchor(
-  point: Point,
-  anchor: { x: number; y: number },
-  radius: number = CANVAS_SETTINGS.ANCHOR_RADIUS
-): boolean {
-  const distance = Math.sqrt(
-    Math.pow(point.x - anchor.x, 2) + Math.pow(point.y - anchor.y, 2)
-  )
-  return distance <= radius
-}
-
-/**
- * 指定座標にあるテキストボックス（アンカー）を検索する
- * @param textBoxes テキストボックスの配列
- * @param point 検索する座標
- * @returns 見つかったテキストボックス、なければnull
- */
-export function findTextBoxAtPoint(
-  textBoxes: TextBox[],
-  point: Point
-): TextBox | null {
-  return (
-    textBoxes.find((textBox) =>
-      isPointNearAnchor(point, { x: textBox.x, y: textBox.y })
-    ) || null
-  )
-}
-
-/**
- * ドラッグ状態から矩形を生成する
- * @param dragState ドラッグ状態
- * @returns 正規化された矩形
- */
-export function createRectFromDrag(dragState: DragState): {
-  x: number
-  y: number
-  width: number
-  height: number
-} {
-  const x = Math.min(dragState.startX, dragState.currentX)
-  const y = Math.min(dragState.startY, dragState.currentY)
-  const width = Math.abs(dragState.currentX - dragState.startX)
-  const height = Math.abs(dragState.currentY - dragState.startY)
-
-  return { x, y, width, height }
-}
-
-/**
- * 新しいテキストボックスのIDを生成する
- * @returns 一意のID文字列
- */
-export function generateTextBoxId(): string {
-  return `textbox-${Date.now()}-${Math.random().toString(36).substr(2, 9)}`
-}
-
-/**
- * クリック位置からアンカーベースのテキストボックスを作成する
- * @param point クリック座標
- * @param text 初期テキスト（デフォルト: 空文字列）
- * @returns 新しいテキストボックス
- */
-export function createAnchorFromClick(
-  point: Point,
-  text: string = ""
-): TextBox {
-  return {
-    id: generateTextBoxId(),
-    x: point.x,
-    y: point.y,
-    text,
-    isSelected: true,
-    anchorDirection: TEXTBOX_SETTINGS.DEFAULT_ANCHOR_DIRECTION,
-    textSize: TEXTBOX_SETTINGS.DEFAULT_TEXT_SIZE,
-  }
-}
-
-/**
- * テキストボックス配列の選択状態を更新する
- * @param textBoxes 既存のテキストボックス配列
- * @param selectedId 選択するテキストボックスのID（nullで全て非選択）
- * @returns 更新されたテキストボックス配列
- */
-export function updateTextBoxSelection(
-  textBoxes: TextBox[],
-  selectedId: string | null
-): TextBox[] {
-  return textBoxes.map((textBox) => ({
-    ...textBox,
-    isSelected: textBox.id === selectedId,
-  }))
-}
-
-/**
- * テキストボックスのテキスト内容を更新する
- * @param textBoxes 既存のテキストボックス配列
- * @param id 更新するテキストボックスのID
- * @param newText 新しいテキスト内容
- * @returns 更新されたテキストボックス配列
- */
-export function updateTextBoxContent(
-  textBoxes: TextBox[],
-  id: string,
-  newText: string
-): TextBox[] {
-  return textBoxes.map((textBox) =>
-    textBox.id === id ? { ...textBox, text: newText } : textBox
-  )
-}
-
-/**
- * テキストボックスを削除する
- * @param textBoxes 既存のテキストボックス配列
- * @param id 削除するテキストボックスのID
- * @returns 更新されたテキストボックス配列
- */
-export function removeTextBox(textBoxes: TextBox[], id: string): TextBox[] {
-  return textBoxes.filter((textBox) => textBox.id !== id)
-}
-
-/**
- * 最小サイズのテキストボックスかどうかを判定する
- * @param rect 矩形オブジェクト
- * @param minWidth 最小幅（デフォルト: 10）
- * @param minHeight 最小高さ（デフォルト: 10）
- * @returns 最小サイズ以上の場合はtrue
- */
-export function isValidTextBoxSize(
-  rect: { width: number; height: number },
-  minWidth: number = 10,
-  minHeight: number = 10
-): boolean {
-  return rect.width >= minWidth && rect.height >= minHeight
-}
-
-/**
- * ドラッグ状態が有効なテキストボックスサイズかどうかを判定する
- * @param dragState ドラッグ状態
- * @param minWidth 最小幅（デフォルト: 10）
- * @param minHeight 最小高さ（デフォルト: 10）
- * @returns 有効なサイズの場合はtrue
- */
-export function isValidDragForTextBox(
-  dragState: DragState,
-  minWidth: number = 10,
-  minHeight: number = 10
-): boolean {
-  const rect = createRectFromDrag(dragState)
-  return isValidTextBoxSize(rect, minWidth, minHeight)
-}
